+++ aquadoggo/src/network/config.rs
@@ -1,7 +1,9 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-use libp2p::Multiaddr;
-use libp2p::{connection_limits::ConnectionLimits, PeerId};
+use std::net::SocketAddr;
+
+use libp2p::connection_limits::ConnectionLimits;
+use libp2p::PeerId;
 
 use crate::AllowList;
 
@@ -23,7 +25,7 @@ pub struct NetworkConfiguration {
     /// with a static IP Address). If you need to connect to nodes with changing, dynamic IP
     /// addresses or even with nodes behind a firewall or NAT, do not use this field but use at
     /// least one relay.
-    pub direct_node_addresses: Vec<Multiaddr>,
+    pub direct_node_addresses: Vec<SocketAddr>,
 
     /// List of peers which are allowed to connect to your node.
     ///
@@ -62,7 +64,7 @@ pub struct NetworkConfiguration {
     /// WARNING: This will potentially expose your IP address on the network. Do only connect to
     /// trusted relays or make sure your IP address is hidden via a VPN or proxy if you're
     /// concerned about leaking your IP.
-    pub relay_addresses: Vec<Multiaddr>,
+    pub relay_addresses: Vec<SocketAddr>,
 
     /// Enable if node should also function as a relay.
     ///