+++ aquadoggo/src/network/utils.rs
@@ -1,7 +1,8 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-use std::net::SocketAddr;
+use std::net::{IpAddr, SocketAddr};
 
+use libp2p::multiaddr::Protocol;
 use libp2p::Multiaddr;
 use regex::Regex;
 
@@ -21,4 +22,14 @@ pub fn to_quic_address(address: &Multiaddr) -> Option<SocketAddr> {
             Some(socket)
         }
     }
+}
+
+pub fn to_multiaddress(socket_address: &SocketAddr) -> Multiaddr {
+    let mut multiaddr = match socket_address.ip() {
+        IpAddr::V4(ip) => Multiaddr::from(Protocol::Ip4(ip)),
+        IpAddr::V6(ip) => Multiaddr::from(Protocol::Ip6(ip)),
+    };
+    multiaddr.push(Protocol::Udp(socket_address.port()));
+    multiaddr.push(Protocol::QuicV1);
+    multiaddr
 }